requirements.txt
pika
psycopg2-binary
paramiko